app_v1/requirements.txt
fastapi
uvicorn
pydantic
SQLAlchemy>=2.0
psycopg[binary]
python-dotenv
streamlit
httpx
prometheus-client
python-json-logger
opensearch-py
# optional but useful for richer uvicorn logging in labs
uvicorn[standard]
